// src/api/__generated__/story/model/batchEpisodeItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export interface BatchEpisodeItem {
  episodeNo: number;
  /**
   * @minLength 0
   * @maxLength 200
   */
  title?: string;
  /**
   * @minLength 0
   * @maxLength 1000
   */
  description: string;
  videoObjectKey: string;
  durationSec: number;
  width: number;
  height: number;
}